software testing: init.s and linker script, make targets for programs
For the time being, using verilator to perform non-uvm tests on the single cycle cpu.  Been having issues running
vsim due to lib.so.6.  Wound up getting this working on a CentOS7 VM but copying the files over is extremely
inconvenient.

init.s currently just sets the stack pointer to 4096 so that compiled programs put the stack pointer in a valid
range of the data memory.

Make targets for the programs are nice but primitive.  Can probably make parameterized functions to just reuse
the logic for each program name.