lsu: refactor to get rid of typedefs in lsu_pkg for buffer entry types
I don't think using separate signals for every field of the load and store queue buffers is particularly perfect, but I opted for this for a few reasons.
Firstly, yosys has limited support for SystemVerilog to begin with, and that does not include pacakges.  I could have used sv2v to convert the project into Verilog before synthesizing with yosys, but I'd rather this project just work in SystemVerilog.  Also I'm sure packages work splendidly with vendor tools like Vivado, but I want this project to not be tied to a vendor tool specifically.  Open source tools like verilator and yosys have been very helpful in my development setup.
Secondly, packages were going to need to be split into a config package and a typedef package, and every other module in the project would need to be refactored (albeit lightly) to use the config package for parameters instead of module parameters.  That means the entire project would break with yosys as described above.
Thirdly, according to ChatGPT, I was going to need to refactor the typedefs to be a struct of arrays instead of an array of structs.  This has a number of benefits, primarily stemming from having access to each field of the buffer in a single array without having to isolate it with a for loop.  This then make rotations, age comparison, and priority encoding more accessible.  The only difference between the typedef and the implementation I'm committing here is that the typedef allows the buffer to be a single port instead of a port for every field in the buffer.  However, I am actually quite fine with modules needing to explicitly state which fields of the buffer they are reading in their port list.  Verbosity is not always the enemy.